Refactored BarcodeGenerator JPG/PNG classes to a more uniform structure
